ITERATION 1 - finalised solution with dependency injection, tests

+++ src/main/java/WordCounter.java
@@ -1,5 +1,10 @@
 public interface WordCounter {
 
-    int countWords(String inputSentence);
+    /**
+     * Takes a user input string and counts all alphabetical words
+     * @param inputSentence user input string
+     * @return number of words
+     */
+    long countWords(String inputSentence);
 
 }
